refactor: extract perspectives routes into dedicated modules
- Create server/src/routes/admin/perspectives.ts for admin CRUD operations
- Create server/src/routes/perspectives.ts for public and user routes
- Remove ~550 lines of inline routes from http.ts
- Add security fixes from code review:
  - Add status parameter validation in user perspectives query
  - Fix author_user_id override vulnerability (always use authenticated user)
  - Add URL scheme validation and 10s timeout for fetch-url endpoint
  - Return 201 status for admin POST (consistency with user route)
- Update admin-tools.ts to use API instead of direct DB queries
