+++ .changeset/perspectives-routes-refactor.md
@@ -1,0 +1,4 @@
+---
+---
+
+Refactor perspectives routes from http.ts into dedicated route files.
