+++ server/src/http.ts
@@ -50,7 +50,8 @@ import { sendAccountLinkedMessage, invalidateMemberContextCache, getAddieBoltRou
 import { createSlackRouter } from "./routes/slack.js";
 import { createWebhooksRouter } from "./routes/webhooks.js";
 import { createWorkOSWebhooksRouter } from "./routes/workos-webhooks.js";
-import { createAdminSlackRouter, createAdminEmailRouter, createAdminFeedsRouter, createAdminNotificationChannelsRouter, createAdminUsersRouter, createAdminSettingsRouter } from "./routes/admin/index.js";
+import { createAdminSlackRouter, createAdminEmailRouter, createAdminFeedsRouter, createAdminNotificationChannelsRouter, createAdminUsersRouter, createAdminSettingsRouter, createAdminPerspectivesRouter } from "./routes/admin/index.js";
+import { createPerspectivesRouter, createUserPerspectivesRouter } from "./routes/perspectives.js";
 import { processFeedsToFetch } from "./addie/services/feed-fetcher.js";
 import { processAlerts } from "./addie/services/industry-alerts.js";
 import { createBillingRouter } from "./routes/billing.js";
@@ -58,11 +59,10 @@ import { createPublicBillingRouter } from "./routes/billing-public.js";
 import { createOrganizationsRouter } from "./routes/organizations.js";
 import { createEventsRouter } from "./routes/events.js";
 import { createLatestRouter } from "./routes/latest.js";
-import { decodeHtmlEntities } from "./utils/html-entities.js";
 import { createCommitteeRouters } from "./routes/committees.js";
 import { sendWelcomeEmail, sendUserSignupEmail, emailDb } from "./notifications/email.js";
 import { emailPrefsDb } from "./db/email-preferences-db.js";
-import { queuePerspectiveLink, processPendingResources, processRssPerspectives, processCommunityArticles } from "./addie/services/content-curator.js";
+import { processPendingResources, processRssPerspectives, processCommunityArticles } from "./addie/services/content-curator.js";
 import { sendCommunityReplies } from "./addie/services/community-articles.js";
 import { sendChannelMessage } from "./slack/client.js";
 import { runTaskReminderJob } from "./addie/jobs/task-reminder.js";
@@ -788,6 +788,14 @@ export class HTTPServer {
     this.app.use('/api/admin/users', adminUsersRouter); // Admin Users: /api/admin/users/*
     const adminSettingsRouter = createAdminSettingsRouter();
     this.app.use('/api/admin/settings', adminSettingsRouter); // Admin Settings: /api/admin/settings/*
+    const adminPerspectivesRouter = createAdminPerspectivesRouter();
+    this.app.use('/api/admin/perspectives', adminPerspectivesRouter); // Admin Perspectives: /api/admin/perspectives/*
+
+    // Mount public and user perspectives routes
+    const perspectivesRouter = createPerspectivesRouter();
+    this.app.use('/api/perspectives', perspectivesRouter); // Public Perspectives: /api/perspectives/*
+    const userPerspectivesRouter = createUserPerspectivesRouter();
+    this.app.use('/api/me/perspectives', userPerspectivesRouter); // User Perspectives: /api/me/perspectives/*
 
     // Mount billing routes (admin)
     const { pageRouter: billingPageRouter, apiRouter: billingApiRouter } = createBillingRouter();
@@ -3237,390 +3245,6 @@ export class HTTPServer {
     });
 
     // ========================================
-    // Perspectives Admin Routes
-    // ========================================
-
-    // GET /api/admin/perspectives - List all perspectives
-    this.app.get('/api/admin/perspectives', requireAuth, requireAdmin, async (req, res) => {
-      try {
-        const pool = getPool();
-        const result = await pool.query(
-          `SELECT * FROM perspectives
-           ORDER BY display_order ASC, published_at DESC NULLS LAST, created_at DESC`
-        );
-
-        res.json(result.rows);
-      } catch (error) {
-        logger.error({ err: error }, 'Get all perspectives error:');
-        res.status(500).json({
-          error: 'Failed to get perspectives',
-          message: error instanceof Error ? error.message : 'Unknown error',
-        });
-      }
-    });
-
-    // GET /api/admin/perspectives/:id - Get single perspective
-    this.app.get('/api/admin/perspectives/:id', requireAuth, requireAdmin, async (req, res) => {
-      try {
-        const { id } = req.params;
-        const pool = getPool();
-        const result = await pool.query(
-          'SELECT * FROM perspectives WHERE id = $1',
-          [id]
-        );
-
-        if (result.rows.length === 0) {
-          return res.status(404).json({
-            error: 'Perspective not found',
-            message: `No perspective found with id ${id}`
-          });
-        }
-
-        res.json(result.rows[0]);
-      } catch (error) {
-        logger.error({ err: error }, 'Get perspective error:');
-        res.status(500).json({
-          error: 'Failed to get perspective',
-          message: error instanceof Error ? error.message : 'Unknown error',
-        });
-      }
-    });
-
-    // POST /api/admin/perspectives/fetch-url - Fetch URL metadata for auto-fill
-    this.app.post('/api/admin/perspectives/fetch-url', requireAuth, requireAdmin, async (req, res) => {
-      try {
-        const { url } = req.body;
-
-        if (!url) {
-          return res.status(400).json({
-            error: 'URL required',
-            message: 'Please provide a URL to fetch'
-          });
-        }
-
-        // Fetch the page
-        const response = await fetch(url, {
-          headers: {
-            'User-Agent': 'Mozilla/5.0 (compatible; AgenticAdvertising/1.0)',
-            'Accept': 'text/html,application/xhtml+xml'
-          },
-          redirect: 'follow'
-        });
-
-        if (!response.ok) {
-          throw new Error(`Failed to fetch URL: ${response.status}`);
-        }
-
-        const html = await response.text();
-
-        // Extract metadata from HTML
-        const titleMatch = html.match(/<title[^>]*>([^<]+)<\/title>/i);
-        const ogTitleMatch = html.match(/<meta[^>]+property=["']og:title["'][^>]+content=["']([^"']+)["']/i)
-          || html.match(/<meta[^>]+content=["']([^"']+)["'][^>]+property=["']og:title["']/i);
-        const ogDescMatch = html.match(/<meta[^>]+property=["']og:description["'][^>]+content=["']([^"']+)["']/i)
-          || html.match(/<meta[^>]+content=["']([^"']+)["'][^>]+property=["']og:description["']/i);
-        const descMatch = html.match(/<meta[^>]+name=["']description["'][^>]+content=["']([^"']+)["']/i)
-          || html.match(/<meta[^>]+content=["']([^"']+)["'][^>]+name=["']description["']/i);
-        const ogSiteMatch = html.match(/<meta[^>]+property=["']og:site_name["'][^>]+content=["']([^"']+)["']/i)
-          || html.match(/<meta[^>]+content=["']([^"']+)["'][^>]+property=["']og:site_name["']/i);
-
-        // Determine title (prefer og:title, then <title>)
-        let title = ogTitleMatch?.[1] || titleMatch?.[1] || '';
-        title = decodeHtmlEntities(title.trim());
-
-        // Determine description (prefer og:description, then meta description)
-        let excerpt = ogDescMatch?.[1] || descMatch?.[1] || '';
-        excerpt = decodeHtmlEntities(excerpt.trim());
-
-        // Site name from og:site_name or parse from URL
-        let site_name = ogSiteMatch?.[1] || '';
-        if (!site_name) {
-          try {
-            const parsedUrl = new URL(url);
-            site_name = parsedUrl.hostname.replace('www.', '');
-            // Capitalize first letter
-            site_name = site_name.charAt(0).toUpperCase() + site_name.slice(1);
-          } catch {
-            // ignore URL parse errors
-          }
-        }
-        site_name = decodeHtmlEntities(site_name);
-
-        res.json({
-          title,
-          excerpt,
-          site_name
-        });
-
-      } catch (error) {
-        logger.error({ err: error }, 'Fetch URL metadata error:');
-        res.status(500).json({
-          error: 'Failed to fetch URL',
-          message: error instanceof Error ? error.message : 'Unknown error'
-        });
-      }
-    });
-
-    // POST /api/admin/perspectives - Create new perspective
-    this.app.post('/api/admin/perspectives', requireAuth, requireAdmin, async (req, res) => {
-      try {
-        const {
-          slug,
-          content_type = 'article',
-          title,
-          subtitle,
-          category,
-          excerpt,
-          content,
-          external_url,
-          external_site_name,
-          author_name,
-          author_title,
-          featured_image_url,
-          status = 'draft',
-          published_at,
-          display_order = 0,
-          tags = [],
-          metadata = {},
-        } = req.body;
-
-        const validContentTypes = ['article', 'link'];
-        const validStatuses = ['draft', 'published', 'archived'];
-
-        if (!slug || !title) {
-          return res.status(400).json({
-            error: 'Missing required fields',
-            message: 'slug and title are required'
-          });
-        }
-
-        if (!validContentTypes.includes(content_type)) {
-          return res.status(400).json({
-            error: 'Invalid content_type',
-            message: 'content_type must be: article or link'
-          });
-        }
-
-        if (!validStatuses.includes(status)) {
-          return res.status(400).json({
-            error: 'Invalid status',
-            message: 'status must be: draft, published, or archived'
-          });
-        }
-
-        // Validate content_type requirements
-        if (content_type === 'link' && !external_url) {
-          return res.status(400).json({
-            error: 'Missing external_url',
-            message: 'external_url is required for link type perspectives'
-          });
-        }
-
-        const pool = getPool();
-        const result = await pool.query(
-          `INSERT INTO perspectives (
-            slug, content_type, title, subtitle, category, excerpt,
-            content, external_url, external_site_name,
-            author_name, author_title, featured_image_url,
-            status, published_at, display_order, tags, metadata
-          )
-          VALUES ($1, $2, $3, $4, $5, $6, $7, $8, $9, $10, $11, $12, $13, $14, $15, $16, $17)
-          RETURNING *`,
-          [
-            slug, content_type, title, subtitle, category, excerpt,
-            content, external_url, external_site_name,
-            author_name, author_title, featured_image_url,
-            status, published_at || null, display_order, tags, metadata
-          ]
-        );
-
-        const perspective = result.rows[0];
-
-        // Queue external links for Addie's knowledge base when published
-        if (perspective.content_type === 'link' && perspective.status === 'published' && perspective.external_url) {
-          queuePerspectiveLink({
-            id: perspective.id,
-            title: perspective.title,
-            external_url: perspective.external_url,
-            category: perspective.category || 'perspective',
-            tags: perspective.tags,
-          }).catch(err => {
-            logger.warn({ err, perspectiveId: perspective.id }, 'Failed to queue perspective link for indexing');
-          });
-        }
-
-        res.json(perspective);
-      } catch (error) {
-        logger.error({ err: error }, 'Create perspective error:');
-        // Check for unique constraint violation
-        if (error instanceof Error && error.message.includes('duplicate key')) {
-          return res.status(400).json({
-            error: 'Slug already exists',
-            message: 'A perspective with this slug already exists'
-          });
-        }
-        res.status(500).json({
-          error: 'Failed to create perspective',
-          message: error instanceof Error ? error.message : 'Unknown error',
-        });
-      }
-    });
-
-    // PUT /api/admin/perspectives/:id - Update perspective
-    this.app.put('/api/admin/perspectives/:id', requireAuth, requireAdmin, async (req, res) => {
-      try {
-        const { id } = req.params;
-        const {
-          slug,
-          content_type,
-          title,
-          subtitle,
-          category,
-          excerpt,
-          content,
-          external_url,
-          external_site_name,
-          author_name,
-          author_title,
-          featured_image_url,
-          status,
-          published_at,
-          display_order,
-          tags,
-          metadata,
-        } = req.body;
-
-        const validContentTypes = ['article', 'link'];
-        const validStatuses = ['draft', 'published', 'archived'];
-
-        if (!slug || !title) {
-          return res.status(400).json({
-            error: 'Missing required fields',
-            message: 'slug and title are required'
-          });
-        }
-
-        if (content_type && !validContentTypes.includes(content_type)) {
-          return res.status(400).json({
-            error: 'Invalid content_type',
-            message: 'content_type must be: article or link'
-          });
-        }
-
-        if (status && !validStatuses.includes(status)) {
-          return res.status(400).json({
-            error: 'Invalid status',
-            message: 'status must be: draft, published, or archived'
-          });
-        }
-
-        // Validate content_type requirements
-        if (content_type === 'link' && !external_url) {
-          return res.status(400).json({
-            error: 'Missing external_url',
-            message: 'external_url is required for link type perspectives'
-          });
-        }
-
-        const pool = getPool();
-        const result = await pool.query(
-          `UPDATE perspectives SET
-            slug = $1,
-            content_type = $2,
-            title = $3,
-            subtitle = $4,
-            category = $5,
-            excerpt = $6,
-            content = $7,
-            external_url = $8,
-            external_site_name = $9,
-            author_name = $10,
-            author_title = $11,
-            featured_image_url = $12,
-            status = $13,
-            published_at = $14,
-            display_order = $15,
-            tags = $16,
-            metadata = $17
-          WHERE id = $18
-          RETURNING *`,
-          [
-            slug, content_type, title, subtitle, category, excerpt,
-            content, external_url, external_site_name,
-            author_name, author_title, featured_image_url,
-            status, published_at || null, display_order, tags, metadata,
-            id
-          ]
-        );
-
-        if (result.rows.length === 0) {
-          return res.status(404).json({
-            error: 'Perspective not found',
-            message: `No perspective found with id ${id}`
-          });
-        }
-
-        const perspective = result.rows[0];
-
-        // Queue external links for indexing when perspective is published
-        if (perspective.content_type === 'link' && perspective.status === 'published' && perspective.external_url) {
-          queuePerspectiveLink({
-            id: perspective.id,
-            title: perspective.title,
-            external_url: perspective.external_url,
-            category: perspective.category || 'perspective',
-            tags: perspective.tags,
-          }).catch(err => {
-            logger.warn({ err, perspectiveId: perspective.id }, 'Failed to queue perspective link for indexing');
-          });
-        }
-
-        res.json(perspective);
-      } catch (error) {
-        logger.error({ err: error }, 'Update perspective error:');
-        // Check for unique constraint violation
-        if (error instanceof Error && error.message.includes('duplicate key')) {
-          return res.status(400).json({
-            error: 'Slug already exists',
-            message: 'A perspective with this slug already exists'
-          });
-        }
-        res.status(500).json({
-          error: 'Failed to update perspective',
-          message: error instanceof Error ? error.message : 'Unknown error',
-        });
-      }
-    });
-
-    // DELETE /api/admin/perspectives/:id - Delete perspective
-    this.app.delete('/api/admin/perspectives/:id', requireAuth, requireAdmin, async (req, res) => {
-      try {
-        const { id } = req.params;
-        const pool = getPool();
-
-        const result = await pool.query(
-          'DELETE FROM perspectives WHERE id = $1 RETURNING id',
-          [id]
-        );
-
-        if (result.rows.length === 0) {
-          return res.status(404).json({
-            error: 'Perspective not found',
-            message: `No perspective found with id ${id}`
-          });
-        }
-
-        res.json({ success: true, deleted: id });
-      } catch (error) {
-        logger.error({ err: error }, 'Delete perspective error:');
-        res.status(500).json({
-          error: 'Failed to delete perspective',
-          message: error instanceof Error ? error.message : 'Unknown error',
-        });
-      }
-    });
-
-    // ========================================
     // Committee Routes (Working Groups, Councils, Chapters)
     // ========================================
 
@@ -3709,169 +3333,6 @@ Disallow: /api/admin/
 `;
       res.set('Content-Type', 'text/plain');
       res.send(robotsTxt);
-    });
-
-    // ========================================
-    // Public Perspectives API Routes
-    // ========================================
-
-    // GET /api/perspectives - List published perspectives (excludes working group posts)
-    this.app.get('/api/perspectives', async (req, res) => {
-      try {
-        const pool = getPool();
-        const result = await pool.query(
-          `SELECT
-            id, slug, content_type, title, subtitle, category, excerpt,
-            external_url, external_site_name,
-            author_name, author_title, featured_image_url,
-            published_at, display_order, tags, like_count
-          FROM perspectives
-          WHERE status = 'published' AND working_group_id IS NULL
-          ORDER BY published_at DESC NULLS LAST`
-        );
-
-        res.json(result.rows);
-      } catch (error) {
-        logger.error({ err: error }, 'Get published perspectives error:');
-        res.status(500).json({
-          error: 'Failed to get perspectives',
-          message: error instanceof Error ? error.message : 'Unknown error',
-        });
-      }
-    });
-
-    // GET /api/perspectives/:slug - Get single published perspective by slug
-    this.app.get('/api/perspectives/:slug', async (req, res) => {
-      try {
-        const { slug } = req.params;
-        const pool = getPool();
-        const result = await pool.query(
-          `SELECT
-            id, slug, content_type, title, subtitle, category, excerpt,
-            content, external_url, external_site_name,
-            author_name, author_title, featured_image_url,
-            published_at, tags, metadata, like_count, updated_at
-          FROM perspectives
-          WHERE slug = $1 AND status = 'published'`,
-          [slug]
-        );
-
-        if (result.rows.length === 0) {
-          return res.status(404).json({
-            error: 'Perspective not found',
-            message: `No published perspective found with slug ${slug}`
-          });
-        }
-
-        res.json(result.rows[0]);
-      } catch (error) {
-        logger.error({ err: error }, 'Get perspective by slug error:');
-        res.status(500).json({
-          error: 'Failed to get perspective',
-          message: error instanceof Error ? error.message : 'Unknown error',
-        });
-      }
-    });
-
-    // POST /api/perspectives/:id/like - Add a like to a perspective
-    this.app.post('/api/perspectives/:id/like', async (req, res) => {
-      try {
-        const { id } = req.params;
-        const { fingerprint } = req.body;
-
-        if (!fingerprint) {
-          return res.status(400).json({
-            error: 'Missing fingerprint',
-            message: 'A fingerprint is required to like a perspective'
-          });
-        }
-
-        const pool = getPool();
-
-        // Get IP hash for rate limiting
-        const ip = req.ip || req.socket.remoteAddress || '';
-        const ipHash = crypto.createHash('sha256').update(ip).digest('hex').substring(0, 64);
-
-        // Check rate limit (max 50 likes per IP per hour)
-        const rateLimitResult = await pool.query(
-          `SELECT COUNT(*) as count FROM perspective_likes
-           WHERE ip_hash = $1 AND created_at > NOW() - INTERVAL '1 hour'`,
-          [ipHash]
-        );
-
-        if (parseInt(rateLimitResult.rows[0].count) >= 50) {
-          return res.status(429).json({
-            error: 'Rate limited',
-            message: 'Too many likes. Please try again later.'
-          });
-        }
-
-        // Insert the like (will fail if already exists due to unique constraint)
-        await pool.query(
-          `INSERT INTO perspective_likes (perspective_id, fingerprint, ip_hash)
-           VALUES ($1, $2, $3)
-           ON CONFLICT (perspective_id, fingerprint) DO NOTHING`,
-          [id, fingerprint, ipHash]
-        );
-
-        // Get updated like count
-        const countResult = await pool.query(
-          `SELECT like_count FROM perspectives WHERE id = $1`,
-          [id]
-        );
-
-        res.json({
-          success: true,
-          like_count: countResult.rows[0]?.like_count || 0
-        });
-      } catch (error) {
-        logger.error({ err: error }, 'Add perspective like error:');
-        res.status(500).json({
-          error: 'Failed to add like',
-          message: error instanceof Error ? error.message : 'Unknown error',
-        });
-      }
-    });
-
-    // DELETE /api/perspectives/:id/like - Remove a like from a perspective
-    this.app.delete('/api/perspectives/:id/like', async (req, res) => {
-      try {
-        const { id } = req.params;
-        const { fingerprint } = req.body;
-
-        if (!fingerprint) {
-          return res.status(400).json({
-            error: 'Missing fingerprint',
-            message: 'A fingerprint is required to unlike a perspective'
-          });
-        }
-
-        const pool = getPool();
-
-        // Delete the like
-        await pool.query(
-          `DELETE FROM perspective_likes
-           WHERE perspective_id = $1 AND fingerprint = $2`,
-          [id, fingerprint]
-        );
-
-        // Get updated like count
-        const countResult = await pool.query(
-          `SELECT like_count FROM perspectives WHERE id = $1`,
-          [id]
-        );
-
-        res.json({
-          success: true,
-          like_count: countResult.rows[0]?.like_count || 0
-        });
-      } catch (error) {
-        logger.error({ err: error }, 'Remove perspective like error:');
-        res.status(500).json({
-          error: 'Failed to remove like',
-          message: error instanceof Error ? error.message : 'Unknown error',
-        });
-      }
     });
 
     // Serve admin pages
@@ -4880,139 +4341,6 @@ Disallow: /api/admin/
         logger.error({ err: error }, 'Get current user error:');
         res.status(500).json({
           error: 'Failed to get user info',
-          message: error instanceof Error ? error.message : 'Unknown error',
-        });
-      }
-    });
-
-    // ========================================
-    // User Perspectives Routes (for Addie CMS tools)
-    // ========================================
-
-    // GET /api/me/perspectives - Get current user's perspectives
-    this.app.get('/api/me/perspectives', requireAuth, async (req, res) => {
-      try {
-        const user = req.user!;
-        const status = req.query.status as string | undefined;
-        const limit = Math.min(parseInt(req.query.limit as string) || 10, 50);
-
-        const pool = getPool();
-
-        let query = `
-          SELECT id, slug, content_type, title, subtitle, category, excerpt,
-                 external_url, external_site_name, author_name, author_title,
-                 status, published_at, created_at, updated_at, tags
-          FROM perspectives
-          WHERE author_user_id = $1 AND working_group_id IS NULL
-        `;
-        const params: (string | number)[] = [user.id];
-
-        if (status && status !== 'all') {
-          query += ` AND status = $${params.length + 1}`;
-          params.push(status);
-        }
-
-        query += ` ORDER BY created_at DESC LIMIT $${params.length + 1}`;
-        params.push(limit);
-
-        const result = await pool.query(query, params);
-        res.json(result.rows);
-      } catch (error) {
-        logger.error({ err: error }, 'GET /api/me/perspectives error');
-        res.status(500).json({
-          error: 'Failed to get perspectives',
-          message: error instanceof Error ? error.message : 'Unknown error',
-        });
-      }
-    });
-
-    // POST /api/me/perspectives - Create a new perspective (as draft)
-    this.app.post('/api/me/perspectives', requireAuth, async (req, res) => {
-      try {
-        const user = req.user!;
-        const {
-          title,
-          content,
-          content_type = 'article',
-          excerpt,
-          external_url,
-          external_site_name,
-          category,
-          tags = [],
-          author_name,
-          author_user_id,
-        } = req.body;
-
-        if (!title) {
-          return res.status(400).json({
-            error: 'Missing required fields',
-            message: 'title is required',
-          });
-        }
-
-        // Validate content_type
-        const validContentTypes = ['article', 'link'];
-        if (!validContentTypes.includes(content_type)) {
-          return res.status(400).json({
-            error: 'Invalid content_type',
-            message: 'content_type must be: article or link',
-          });
-        }
-
-        // Validate content_type requirements
-        if (content_type === 'link' && !external_url) {
-          return res.status(400).json({
-            error: 'Missing external_url',
-            message: 'external_url is required for link type perspectives',
-          });
-        }
-
-        if (content_type === 'article' && !content) {
-          return res.status(400).json({
-            error: 'Missing content',
-            message: 'content is required for article type perspectives',
-          });
-        }
-
-        // Generate slug from title
-        const baseSlug = title
-          .toLowerCase()
-          .replace(/[^a-z0-9\s-]/g, '')
-          .replace(/\s+/g, '-')
-          .substring(0, 100);
-
-        // Add timestamp suffix to make unique
-        const slug = `${baseSlug}-${Date.now().toString(36)}`;
-
-        // Use provided author info or derive from user
-        const finalAuthorName = author_name ||
-          (user.firstName && user.lastName
-            ? `${user.firstName} ${user.lastName}`
-            : user.email?.split('@')[0] || 'Anonymous');
-        const finalAuthorUserId = author_user_id || user.id;
-
-        const pool = getPool();
-        const result = await pool.query(
-          `INSERT INTO perspectives (
-            slug, content_type, title, content, excerpt,
-            external_url, external_site_name, category, tags,
-            author_name, author_user_id, status
-          )
-          VALUES ($1, $2, $3, $4, $5, $6, $7, $8, $9, $10, $11, 'draft')
-          RETURNING *`,
-          [
-            slug, content_type, title, content, excerpt,
-            external_url, external_site_name, category, tags,
-            finalAuthorName, finalAuthorUserId,
-          ]
-        );
-
-        logger.info({ perspectiveId: result.rows[0].id, userId: user.id, title }, 'User created perspective draft');
-        res.status(201).json(result.rows[0]);
-      } catch (error) {
-        logger.error({ err: error }, 'POST /api/me/perspectives error');
-        res.status(500).json({
-          error: 'Failed to create perspective',
           message: error instanceof Error ? error.message : 'Unknown error',
         });
       }
